CreateParser

Description

Documentation

data CreateExpr

CreateExpr is a constructor for Create Queries

Constructors

Create

Fields

iTname :: ValueExpr

iTname is name of the table

iColLists :: [(ValueExpr, Datatype)]

iColLists contains list of column names and their dataype values

createExpr :: Parser CreateExpr

createExpr aggerates all parsers to parse CREATE query

makeCreate :: CreateExpr

makeCreate creates empty CreateExpr object

evaluateCreate :: Either ParseError CreateExpr -> Maybe Database -> Maybe Database

evaluateCreate evaluates parsed CreateExpr to create a new table and add columns to it. First argument is parsed CreateExpr Second argument is databse instance