Funcs

Description

Documentation

regularParse :: Parser a -> String -> Either ParseError a

The regularParse parses a expression regularlly

parseWithEof :: Parser a -> String -> Either ParseError a

The parseWithEof is a wrapper that throws error if you haven’t consumed all the input

parseWithLeftOver :: Parser a -> String -> Either ParseError (a, String)

The parseWithLeftOver is wrapper that tell you what was not consumed from the input by input parser

parseWithWSEof :: Parser a -> String -> Either ParseError a

The parseWithWSEof wrapper allows parsing when input constains and leading whitespaces

lexeme :: Parser a -> Parser a

The lexeme is wrapper parser that eats up any following whitespaces after parsing.

integer :: Parser Integer

The integer parses a integer constant

identifier :: Parser String

The identifier parses a identifier string

symbol :: String -> Parser String

The symbol parses a single character symbol or operator It takes string as arguments and comapres it with parsed string.

comma :: Parser Char

The comma parses comma

openParen :: Parser Char

The openParen parses Opening paranthesis

closeParen :: Parser Char

The closeParen parses closing paranthesis

keyword :: String -> Parser String

The keyowrd parses a occurence of string given as input It takes a string to pe parsed as only input

parens :: Parser a -> Parser a

The parens function parses anything between opening and closing paranthesis

identifierBlacklist :: [String] -> Parser String

The identifierBlacklist parses only that identifiers that are not in blacklist It takes a blacklist of type List as an argument

whitespace :: Parser ()

The whitespace parses any number of any type of whiespaces

keyword_ :: String -> Parser ()

The keyword_ parses keywords and ignores them

symbol_ :: String -> Parser ()

The symbol_ parses symbol and ignores them

commaSep1 :: Parser a -> Parser [a]

The commaSep1 parses comma seperated items and returns a list

stringToken :: Parser String

The stringToken parses a string literal in single quotes

boolToken :: Parser Bool

The boolToken parses boolean constants True and False