QueryParser

Description

Documentation

data QueryExpr

QueryExpr represents a SELECT statement

Constructors

Select

Fields

qeSelectList :: [(ValueExpr, Maybe String)]

qeSelectList contains list of selected columns

qefromClause :: ValueExpr

qeFromClause contains table name

qeWhere :: Maybe ValueExpr

qeWhere contains where clause expression

qeGroupBy :: [ValueExpr]

qeGroupBy contains list of groupby expresions

qeHaving :: Maybe ValueExpr

qeHaving contains havingby expression

qeOrderBy :: [ValueExpr]

qeOrderBy contains list of orderby clause expressions

queryExpr :: Parser QueryExpr

queryExpr aggregates all parsers to parse a SELECT statement

makeSelect :: QueryExpr

makeSelect makes a empty QueryExpr object

evaluateQuery :: Either ParseError QueryExpr -> Maybe Database -> [[(Int, String, Datatype, String)]]

evaluateQuery evaluates a query expression First argument is parsed query expression Second argument is Maybe Database instance