SQLParser

Description

Documentation

data SQLExpr

SQLExpr is a wrapping type constructor for Select, Insert and Delete statements. 'SELECT QueryExpr' contains

Constructors

SELECT QueryExpr
Parses a SELECT query

INSERT InsertExpr
Parses a INSERT Query

CREATE CreateExpr
Parses a CREATE Query

DELETE DeleteExpr
Parses a DELETE Query

data ResType

ResType is a wrapping type constructor for the response types of each of the above statements.

Constructors

DB (Maybe Database)
Database instance for insert, create and delete queries

OUT [[(Int, String, Datatype, String)]]
Return object for SELECT queries

ERROR String

sqlExpr :: Parser SQLExpr

sqlExpr describes an SQL expression.

evaluateSQL :: Either ParseError SQLExpr -> ResType -> ResType

evaluateSQL evaluates each Expression using individual case statements

getHeaders :: [(Int, String, Datatype, String)] -> String

getHeader returns string of headers of table First argument is list of headers

getRow :: [(Int, String, Datatype, String)] -> String

getHeader returns string of depicting one entire row First argument is a row

getRows :: [[(Int, String, Datatype, String)]] -> IO ()

getRows prints the rows which are the output of select query The argument is output of select parsed query

queryPrinter :: ResType -> IO ()

queryPrinter prints the Select query output in a tabular fashion The argument is output of select parsed query